Updated issues with reassignment and restructured the nodes to be formatted more universally. Only thing missing is comparison of doubles and strings, which shouldn't be terrible to add. Not something that is required and being tested though, so not a worry.
