+++ Parser/Tree_Nodes/b_stmt.c
@@ -52,6 +52,9 @@ b_stmt *create_b_stmt(GArray *token_stream, unsigned long index, unsigned long *
                         }
                         break;
 
+                    case jf_str:
+                    case jf_double:
+                    case jf_int:
                     case jint:
                     case jdouble:
                     case jstring:
@@ -78,175 +81,14 @@ b_stmt *create_b_stmt(GArray *token_stream, unsigned long index, unsigned long *
         }
             break;
         case t_while:
-            (*next)++;
-            // Load next token
-            Token *check = &g_array_index(token_stream, Token, *next);
-            //Is it a parend
-            if (check->type == t_start_paren) {
-                (*next)++;
-            } else {
-                fprintf(stderr,
-                        "Syntax Error: missing start parend in while statement");
-                exit(-1);
-            }
-
-            // Grab expression to evaluate
-            i_expr *while_conditional = create_i_expr(token_stream, *next, next);
-
-            check = &g_array_index(token_stream, Token, *next);
-
-            // Is there an end parend and starting bracket for b_stmt?
-            if (check->type == t_end_paren) {
-                (*next)++;
-                check = &g_array_index(token_stream, Token, *next);
-            } else {
-                fprintf(stderr,
-                        "Syntax Error: missing end parend in while statement");
-                exit(-1);
-            }
-
-            // Is there an opening bracket
-            if (check->type == t_start_bracket) {
-                (*next)++;
-            } else {
-                fprintf(stderr,
-                        "Syntax Error: missing start bracket in while statement");
-                exit(-1);
-            }
-
-            // Make B_stmt_list for body
-            b_stmt_list *while_body = create_b_stmt_list(token_stream, *next, next);
-            check = &g_array_index(token_stream, Token, (*next));
-            if (check->type == t_end_bracket) {
-                (*next)++;
-                new_statement->whileLoop = create_while_node(while_conditional, while_body);
-                break;
-            } else {
-                fprintf(stderr, "Syntax Error: Missing } at end of while statement");
-                exit(-1);
-            }
-
+            new_statement->whileLoop = create_while_node(token_stream, (*next), next);
+            break;
         case t_for:
-            (*next)++;
-            check = &g_array_index(token_stream, Token, *next);
-            if (check->type == t_start_paren) {
-                (*next)++;
-            } else {
-                fprintf(stderr,
-                        "Syntax Error: missing start parend in for loop");
-                exit(-1);
-            }
-            // Grab first assignment
-            asmt *for_initialize = create_asmt(token_stream, *next, next);
-
-            // Grab the conditional statement next
-            i_expr *for_conditional = create_i_expr(token_stream, *next, next);
-            //Check next char
-            check = &g_array_index(token_stream, Token, *next);
-            if (check->type == t_end_stmt) {
-                (*next)++;
-            } else {
-                fprintf(stderr,
-                        "Syntax Error: missing second ; in for loop");
-                exit(-1);
-            }
-
-            // Grab the incrementer reassignment function
-            r_asmt *for_incrementer = create_r_asmt(token_stream, *next, next);
-
-            // Check next parend
-            check = &g_array_index(token_stream, Token, *next);
-            if (check->type == t_end_paren) {
-                (*next)++;
-            } else {
-                fprintf(stderr,
-                        "Syntax Error: missing end parend on for loop");
-                exit(-1);
-            }
-
-            //Check for closing bracket
-            check = &g_array_index(token_stream, Token, *next);
-            if (check->type == t_start_bracket) {
-                (*next)++;
-            } else {
-                fprintf(stderr,
-                        "Syntax Error: missing start bracket in for loop");
-                exit(-1);
-            }
-            b_stmt_list *for_body = create_b_stmt_list(token_stream, *next, next);
-            check = &g_array_index(token_stream, Token, (*next));
-            if (check->type == t_end_bracket) {
-                (*next)++;
-                new_statement->forLoop = create_for_node(for_initialize, for_conditional, for_incrementer, for_body);
-                break;
-            } else {
-                fprintf(stderr, "Syntax Error: Missing } at end of for statement");
-                exit(-1);
-            }
+            new_statement->forLoop = create_for_node(token_stream, (*next), next);
+            break;
         case t_if:
-            (*next)++;
-            check = &g_array_index(token_stream, Token, *next);
-            if (check->type == t_start_paren) {
-                (*next)++;
-            } else {
-                fprintf(stderr,
-                        "Syntax Error: missing start parend in if statement");
-                exit(-1);
-            }
-            // Grab expression to evaluate
-            i_expr *if_conditional = create_i_expr(token_stream, *next, next);
-
-            check = &g_array_index(token_stream, Token, *next);
-            // Is there an end parend and starting bracket for b_stmt?
-            if (check->type == t_end_paren) {
-                (*next)++;
-                check = &g_array_index(token_stream, Token, *next);
-            } else {
-                fprintf(stderr,
-                        "Syntax Error: missing end parend in if statement");
-                exit(-1);
-            }
-            if (check->type == t_start_bracket) {
-                (*next)++;
-            } else {
-                fprintf(stderr,
-                        "Syntax Error: missing start bracket in if statement");
-                exit(-1);
-            }
-
-            // Make B_stmt_list for true branch
-            b_stmt_list *true_path = create_b_stmt_list(token_stream, *next, next);
-            check = &g_array_index(token_stream, Token, *next);
-            if (check->type == t_end_bracket) {
-                (*next)++;
-                new_statement->ifBlock = create_if_node(if_conditional, true_path, NULL);
-                check = &g_array_index(token_stream, Token, *next);
-                if (check->type == t_else) {
-                    (*next)++;
-                    check = &g_array_index(token_stream, Token, *next);
-                    if (check->type != t_start_bracket) {
-                        fprintf(stderr,
-                                "Syntax Error: missing start bracket in else statement");
-                        exit(-1);
-                    }
-                    (*next)++;
-                    b_stmt_list *temp = create_b_stmt_list(token_stream, *next, next);
-                    check = &g_array_index(token_stream, Token, *next);
-                    if (check->type == t_end_bracket) {
-                        (*next)++;
-                        new_statement->ifBlock->b_false = temp;
-                        break;
-                    } else {
-                        fprintf(stderr, "Syntax Error: Missing } at end of else statement");
-                        exit(-1);
-
-                    }
-                }
-                break;
-            } else {
-                fprintf(stderr, "Syntax Error: Missing } at end of if statement");
-                exit(-1);
-            }
+            new_statement->ifBlock = create_if_node(token_stream, (*next), next);
+            break;
         default:
             // Unexpected token when creating Statement, not function call, assignment, or expression
             fprintf(stderr, "Syntax Error: Unexpected Token when creating b_statement %s", curToken->data->str);
@@ -285,6 +127,22 @@ GString *b_stmt_to_json(b_stmt *statement) {
         } else {
             g_string_append(retVal, "null");
         }
+        g_string_append(retVal, ", \"For loop\": ");
+        if (statement->forLoop != NULL) {
+            GString *child = for_node_to_json(statement->forLoop);
+            g_string_append(retVal, child->str);
+            g_string_free(child, TRUE);
+        } else {
+            g_string_append(retVal, "null");
+        }
+        g_string_append(retVal, ", \"While loop\": ");
+        if (statement->forLoop != NULL) {
+            GString *child = while_node_to_json(statement->whileLoop);
+            g_string_append(retVal, child->str);
+            g_string_free(child, TRUE);
+        } else {
+            g_string_append(retVal, "null");
+        }
         g_string_append(retVal, ", \"If Statement\": ");
         if (statement->ifBlock != NULL) {
             GString *child = if_node_to_json(statement->ifBlock);
@@ -293,22 +151,7 @@ GString *b_stmt_to_json(b_stmt *statement) {
         } else {
             g_string_append(retVal, "null");
         }
-        g_string_append(retVal, ", \"While Statement\": ");
-        if (statement->whileLoop != NULL) {
-            GString *child = while_node_to_json(statement->whileLoop);
-            g_string_append(retVal, child->str);
-            g_string_free(child, TRUE);
-        } else {
-            g_string_append(retVal, "null");
-        }
-        g_string_append(retVal, ", \"For Statement\": ");
-        if (statement->forLoop != NULL) {
-            GString *child = for_node_to_json(statement->forLoop);
-            g_string_append(retVal, child->str);
-            g_string_free(child, TRUE);
-        } else {
-            g_string_append(retVal, "null");
-        }
+
         g_string_append_c(retVal, '}');
     } else {
         g_string_append(retVal, "null");
@@ -329,6 +172,12 @@ void destroy_b_stmt(b_stmt *statement) {
         }
         if (statement->ifBlock != NULL) {
             destroy_if_node(statement->ifBlock);
+        }
+        if (statement->whileLoop != NULL) {
+            destroy_while_node(statement->whileLoop);
+        }
+        if (statement->forLoop != NULL) {
+            destroy_for_node(statement->forLoop);
         }
         free(statement);
     }
