+++ Parser/Tree_Nodes/while_node.c
@@ -23,15 +23,57 @@
  * @param next - Lookahead pointer
  * @return new While node with a i_expr conditional and b_stmt_list
  */
-while_node *create_while_node(i_expr *conditional, b_stmt_list *body) {
+while_node *create_while_node(GArray *token_stream, unsigned long index, unsigned long *next) {
     while_node *retVal = NULL;
 
     // Make new While node
     retVal = calloc(1, sizeof(while_node));
-    retVal->body = body;
-    retVal->conditional = conditional;
-    //Return the value
-    return retVal;
+    (*next)++;
+    // Load next token
+    Token *check = &g_array_index(token_stream, Token, *next);
+    //Is it a parend
+    if (check->type == t_start_paren) {
+        (*next)++;
+    } else {
+        fprintf(stderr,
+                "Syntax Error: missing start parend in while statement");
+        exit(-1);
+    }
+
+    // Grab expression to evaluate
+    retVal->conditional = create_i_expr(token_stream, *next, next);
+
+    check = &g_array_index(token_stream, Token, *next);
+
+    // Is there an end parend and starting bracket for b_stmt?
+    if (check->type == t_end_paren) {
+        (*next)++;
+        check = &g_array_index(token_stream, Token, *next);
+    } else {
+        fprintf(stderr,
+                "Syntax Error: missing end parend in while statement");
+        exit(-1);
+    }
+
+    // Is there an opening bracket
+    if (check->type == t_start_bracket) {
+        (*next)++;
+    } else {
+        fprintf(stderr,
+                "Syntax Error: missing start bracket in while statement");
+        exit(-1);
+    }
+
+    // Make B_stmt_list for body
+    retVal->body = create_b_stmt_list(token_stream, *next, next);
+    check = &g_array_index(token_stream, Token, (*next));
+    if (check->type == t_end_bracket) {
+        (*next)++;
+        return retVal;
+    } else {
+        fprintf(stderr, "Syntax Error: Missing } at end of while statement");
+        exit(-1);
+    }
 }
 
 /**
